Move test_workflow_apply to test_workflow.py
